Simplify 2048 dev
